pyproject.toml
[project]
name = "personify"
version = "0.1.0"
description = "personify using crewAI"
authors = [{ name = "Hashirscam", email = "perfactorial@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.140.0,<1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
personify = "personify.main:run"
run_crew = "personify.main:run"
train = "personify.main:train"
replay = "personify.main:replay"
test = "personify.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
